+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 ##### Fixed
 
 - Buttons in search now have a pointer cursor
+- Pokemons page always loads on top, not on saved position
 
 ---