+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - Transition on pokemon details page... again
 - Scroll to top button bottom margin when navigating away from a page with footer
 - Search on mobile
+- Undefined characteristics for species
 
 ---