+++ CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ---
 
-## [Unreleased](https://github.com/nashaguayo/pokedex/compare/V0.2.2...develop) (dd-mm-yyyy)
+## [Unreleased](https://github.com/nashaguayo/pokedex/compare/V0.2.3...develop) (dd-mm-yyyy)
+
+---
+
+## [V0.2.3](https://github.com/nashaguayo/pokedex/compare/V0.2.2...V0.2.3) (23-07-2023)
 
 ##### Added